creates new class to handle time delay estimates between multi dimension timeseries
